+++ app/panel/components/BuildingBlocks/index.js
@@ -21,8 +21,6 @@ import GhosteryFeature from './GhosteryFeature';
 import NotScanned from './NotScanned';
 import PauseButton from './PauseButton';
 import ToggleSlider from './ToggleSlider';
-import RadioButtonGroup from './RadioButtonGroup';
-import RadioButton from './RadioButton';
 import ModalExitButton from './ModalExitButton';
 
 export {
@@ -33,7 +31,5 @@ export {
 	NotScanned,
 	PauseButton,
 	ToggleSlider,
-	RadioButtonGroup,
-	RadioButton,
 	ModalExitButton
 };
