+++ app/shared-components/RadioButtonGroup/RadioButtonGroup.jsx
@@ -13,7 +13,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import RadioButton from './RadioButton';
+import RadioButton from '../RadioButton';
 
 /**
  * @class Implements a radio button group
